+++ src/services/price/services/pyth-oracle.service.ts
@@ -56,6 +56,7 @@ export const fetchPythOracleData = async (
     pythPushKaminosBanks,
     driftPythPullBanks,
     solendPythPullBanks,
+    juplendPythPullBanks,
   } = categorizePythBanks(banks);
 
   if (
@@ -63,7 +64,8 @@ export const fetchPythOracleData = async (
     !pythStakedCollateralBanks.length &&
     !pythPushKaminosBanks.length &&
     !driftPythPullBanks.length &&
-    !solendPythPullBanks.length
+    !solendPythPullBanks.length &&
+    !juplendPythPullBanks.length
   ) {
     // Return empty structures when there are no banks to process
     return {
@@ -100,6 +102,7 @@ export const fetchPythOracleData = async (
     ...pythPushKaminosBanks,
     ...driftPythPullBanks,
     ...solendPythPullBanks,
+    ...juplendPythPullBanks,
   ];
   const pythOracleKeys = extractPythOracleKeys(combinedPythBanks);