+++ src/types.ts
@@ -1,5 +1,6 @@
 import { PublicKey, Transaction, VersionedTransaction } from "@solana/web3.js";
 import { Program as AnchorProgram, AnchorProvider, Idl } from "@coral-xyz/anchor";
+import BN from "bn.js";
 
 import { MarginfiIdlType } from "./idl";
 import { Bank } from "./models/bank";
@@ -15,6 +16,14 @@ import {
   DriftUserStatsJSON,
   FarmStateJSON,
   FarmStateRaw,
+  JupLendingRewardsRateModel,
+  JupLendingRewardsRateModelJSON,
+  JupLendingState,
+  JupLendingStateJSON,
+  JupRateModel,
+  JupRateModelJSON,
+  JupTokenReserve,
+  JupTokenReserveJSON,
   ObligationJSON,
   ObligationRaw,
   ReserveJSON,
@@ -93,6 +102,13 @@ export type BankIntegrationMetadata = {
     userRewards: DriftRewards[];
     userStatsState?: DriftUserStats;
   };
+  jupLendStates?: {
+    jupLendingState: JupLendingState;
+    jupTokenReserveState: JupTokenReserve;
+    jupRewardsRateModel: JupLendingRewardsRateModel | null;
+    jupRateModel: JupRateModel | null;
+    fTokenTotalSupply: BN;
+  };
 };
 
 export type BankIntegrationMetadataDto = {
@@ -106,6 +122,13 @@ export type BankIntegrationMetadataDto = {
     userState: DriftUserJSON;
     userRewards: DriftRewardsJSON[];
     userStatsState?: DriftUserStatsJSON;
+  };
+  jupLendStates?: {
+    jupLendingState: JupLendingStateJSON;
+    jupTokenReserveState: JupTokenReserveJSON;
+    jupRewardsRateModel: JupLendingRewardsRateModelJSON | null;
+    jupRateModel: JupRateModelJSON | null;
+    fTokenTotalSupply: string;
   };
 };